+++ internal/pkg/image/unpacker/squashfs.go
@@ -165,7 +165,15 @@ func (s *Squashfs) extract(files []string, reader io.Reader, dest string) (err e
 		cmd.Stdin = reader
 	}
 
-	if o, err := cmd.CombinedOutput(); err != nil {
+	o, err := cmd.CombinedOutput()
+
+	if os.Getenv("SINGULARITY_DEBUG") != "" {
+		sylog.Debugf("*** BEGIN WRAPPED UNSQUASHFS OUTPUT ***")
+		sylog.Debugf(string(o))
+		sylog.Debugf("*** END WRAPPED UNSQUASHFS OUTPUT ***")
+	}
+
+	if err != nil {
 		return fmt.Errorf("extract command failed: %s: %s", string(o), err)
 	}